Split Accuracy into grayLevelPrecision and sizePrecision
In searchDotInArea(), now we search also dots in the bounding dots of badDots

In isValid(): Now we approximate the dot by an ellipse and try to close it between a smaller and a bigger one with the same parameters. If we don,t success, the dot isn,t valid.

Correction of a bug in searchDotInArea() : the algorithm could fall in an infinite loop.