Before, vpMbtHiddenFace.* and vpMbtKltHiddenFace.* had two classes. Now it is divided in two files : one manages a list of polygons (vpMbhiddenFaces.h)  and the other represents a polygon (vpMbtPolygon.* and vpMbtKltPolygon.*).
Moreover, we introduced a polygon visibility test based on Ogre3D. This test is only available when Ogre3D is detected.
Use setOgreVisibilityTest(bool) to enable/disable this functionnality.
The visibility test is based on RayCasting. A face is visilble if its center of gravity is visible. 

Warning, before Ogre3D 1.8.1 it opens a render window, that displays the visible faces. Since 1.8.1 this window is hidden.