Fix [17972] Ogre examples are not working with Fedora 20
On Fedora 20 when ogre-devel and ogre-samples packages are installed, AROgre and
AROgreBasic examples produce a segfault when the folowing line in resource.cf.in
is commented:
'#FileSystem=@OGRE_MEDIA_DIR@/materials/programs'
Here we check if @OGRE_MEDIA_DIR@/materials/programs forder is empty.
If empty, we comment the line to avoid lintian warnings.
If not empty we remove the comment to use the available programs.