Refactor vpPylonGrabber into a factory pattern
It's frustrating to cope Basler GigE and USB cameras in a single source
code base for the reason that they are using different attribute names
and enumeration definitions. We are refactoring it into an abstract
base class with instances created by a factory singleton. GigE and USB3
cameras are supported by vpPylonGrabber subclasses.