Fix vpMomentAlpha computation
The computation based on `atan()` actually generates orientation angle
in [-pi/4, pi/4]. And the result is either the major axis or the minor
axis which is not specified. This patch will always give out the angle
of the major axis. A span of [-pi/2, pi/2] is achieved by considering
both sine and cosine of 2*alpha.

References

http://homepages.inf.ed.ac.uk/rbf/CVonline/LOCAL_COPIES/OWENS/LECT2/node3.html